requirements.txt
streamlit
pymongo
dnspython
pandas
matplotlib
python-dateutil
